[INJIMOB-3497]: Moving digest mutlibase to Utils
