+++ kotlin/injivcrenderer/src/commonMain/kotlin/io/mosip/injivcrenderer/utils/Utils.kt
@@ -1,7 +1,6 @@
 package io.mosip.injivcrenderer.utils
 
 import com.fasterxml.jackson.databind.JsonNode
-import io.mosip.injivcrenderer.DigestMultibaseHelper
 import io.mosip.injivcrenderer.constants.Constants.DIGEST_MULTIBASE
 import io.mosip.injivcrenderer.networkManager.NetworkManager
 import io.mosip.injivcrenderer.constants.Constants.ID
@@ -9,12 +8,16 @@ import io.mosip.injivcrenderer.constants.Constants.QR_CODE_PLACEHOLDER
 import io.mosip.injivcrenderer.constants.Constants.QR_IMAGE_PREFIX
 import io.mosip.injivcrenderer.constants.Constants.RENDER_METHOD
 import io.mosip.injivcrenderer.constants.Constants.RENDER_SUITE
+import io.mosip.injivcrenderer.constants.Constants.SHA_256
 import io.mosip.injivcrenderer.constants.Constants.SVG_MUSTACHE
 import io.mosip.injivcrenderer.constants.Constants.TEMPLATE
 import io.mosip.injivcrenderer.constants.Constants.TEMPLATE_RENDER_METHOD
 import io.mosip.injivcrenderer.constants.Constants.TYPE
 import io.mosip.injivcrenderer.exceptions.VcRendererExceptions
 import io.mosip.injivcrenderer.qrCode.QrCodeGenerator
+import java.security.MessageDigest
+import kotlin.io.encoding.Base64
+import kotlin.io.encoding.ExperimentalEncodingApi
 
 class Utils(private val traceabilityId: String) {
     private val className = Utils::class.simpleName
@@ -35,7 +38,7 @@ class Utils(private val traceabilityId: String) {
         val digestMultibase = templateValue.path(DIGEST_MULTIBASE).asText(null)
 
         var rawSvg = NetworkManager(traceabilityId).fetchSvgAsText(templateId)
-        if (digestMultibase != null && !DigestMultibaseHelper(traceabilityId).validateDigestMultibase(rawSvg, digestMultibase)) {
+        if (digestMultibase != null && !validateDigestMultibase(rawSvg, digestMultibase)) {
             throw VcRendererExceptions.MultibaseValidationException(
                 traceabilityId = traceabilityId,
                 className = className,
@@ -112,6 +115,31 @@ class Utils(private val traceabilityId: String) {
                 className
             )
         }
+    }
+
+    fun validateDigestMultibase(svgString: String, digestMultibase: String): Boolean {
+        if (!digestMultibase.startsWith("u")) throw VcRendererExceptions.MultibaseValidationException(traceabilityId, className, "digestMultibase must start with 'u'")
+        val encodedPart = digestMultibase.substring(1)
+
+        val decoded = base64UrlNoPadDecode(encodedPart)
+        if (decoded.size != 34)
+            throw VcRendererExceptions.MultibaseValidationException(traceabilityId, className, "Invalid multihash length")
+        if (decoded[0] != 0x12.toByte() || decoded[1] != 0x20.toByte())
+            throw VcRendererExceptions.MultibaseValidationException(traceabilityId, className, "Unsupported multihash prefix")
+
+        val expectedHash = decoded.copyOfRange(2, 34)
+        val actualHash = MessageDigest.getInstance(SHA_256).digest(svgString.toByteArray(Charsets.UTF_8))
+
+        return actualHash.contentEquals(expectedHash)
+    }
+
+    @OptIn(ExperimentalEncodingApi::class)
+    private fun base64UrlNoPadDecode(input: String): ByteArray {
+        val standardBase64 = input
+            .replace('-', '+')
+            .replace('_', '/')
+            .padEnd(input.length + (4 - input.length % 4) % 4, '=')
+        return Base64.decode(standardBase64)
     }
