+++ kotlin/injivcrenderer/src/commonTest/kotlin/io/mosip/injivcrenderer/utils/UtilTest.kt
@@ -4,14 +4,13 @@ import kotlin.test.Test
 import kotlin.test.assertTrue
 import kotlin.test.assertFalse
 import kotlin.test.assertFailsWith
-import io.mosip.injivcrenderer.DigestMultibaseHelper
 import io.mosip.injivcrenderer.TestUtils
 import io.mosip.injivcrenderer.exceptions.VcRendererExceptions
 
-class DigestMultibaseHelperTest {
+class UtilsTest {
 
     private val traceId = "test-trace-id"
-    private val helper = DigestMultibaseHelper(traceId)
+    private val helper = Utils(traceId)
     private val svgSample = """<svg>Email: {{/credentialSubject/email}}, Mobile: {{/credentialSubject/mobile}}</svg>"""
 
     @Test
